+++ CHANGELOG.md
@@ -5,4 +5,6 @@
 - LibUI general control and window support
 - LibUI listener support
 - LibUI property support
+- girb (Glimmer IRB)
 - Support examples/basic_window.rb
+- Support examples/basic_button.rb